Move Graph.svelte
